requirements.txt
Flask==2.2.3
Pillow==9.4.0
fpdf==1.7.2
